pyproject.toml
[tool.poetry]
name = "abctk.utils.comparative"
version = "1.0.0"
description = ""
authors = ["Nori Hayashi <net@hayashi-lin.net>"]
readme = "README.md"
packages = [
    {include = "abctk/utils/comparative"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
typer = "^0.9.0"
ruamel-yaml = "^0.18.5"
lxml = "^5.0.1"
tqdm = "^4.66.1"
pandas = "^2.1.4"
abctk-obj = {git = "https://github.com/ABCTreebank/abctk.obj"}

[tool.poetry.scripts]
"abctk.utils.comparative" = "abctk.utils.comparative.cli:app"

[tool.poetry.group.dev.dependencies]
nuitka = "^1.9.6"
pyinstaller = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
